chatbot-backend/requirements.txt
Flask==2.2.2
Jinja2==3.1.2
gunicorn==21.2.0
line-bot-sdk==2.2.1